Testing changes to make TDS more broadly applicable by moving to tds.py
