+++ impacket/examples/ntlmrelayx/clients/mssqlrelayclient.py
@@ -21,15 +21,14 @@
 #   [ ] Handle SQL Authentication
 #
 
-import ssl
 import struct
 import random
 import string
 
 from impacket import LOG
 from impacket.examples.ntlmrelayx.clients import ProtocolClient
-from impacket.tds import MSSQL, DummyPrint, TDS_ENCRYPT_REQ, TDS_ENCRYPT_OFF, TDS_ENCRYPT_ON, TDS_ENCRYPT_NOT_SUP, \
-    TDS_ENCRYPT_STRICT, TDS_PRE_LOGIN, TDS_LOGIN, TDS_INIT_LANG_FATAL, TDS_ODBC_ON, \
+from impacket.tds import MSSQL, DummyPrint, TDS_ENCRYPT_OFF, TDS_ENCRYPT_REQ, TDS_ENCRYPT_ON, \
+    TDS_ENCRYPT_STRICT, TDS_LOGIN, TDS_INIT_LANG_FATAL, TDS_ODBC_ON, \
     TDS_INTEGRATED_SECURITY_ON, TDS_LOGIN7, TDS_SSPI, TDS_LOGINACK_TOKEN
 from impacket.ntlm import NTLMAuthChallenge
 from impacket.nt_errors import STATUS_SUCCESS, STATUS_ACCESS_DENIED
@@ -42,40 +41,8 @@ class MYMSSQL(MSSQL):
         MSSQL.__init__(self, address, port, rowsPrinter)
         self.resp = None
         self.sessionData = {}
-        self.tds8 = False
-
-    def socketRecv(self, bufsize):
-        """Override to detect closed connections instead of looping forever in recvTDS.
-
-        The base class recvTDS has a `while data == b""` loop. If the server closes
-        the connection (e.g. Force Strict Encryption rejects plain TDS), socket.recv()
-        returns b"" immediately and the loop spins forever. This override raises
-        ConnectionError so the caller can fall back to TDS 8.0.
-        """
-        if self.tlsSocket is None:
-            data = self.socket.recv(bufsize)
-            if not data:
-                raise ConnectionError("Server closed connection")
-            return data
-        else:
-            return self.tls_recv(bufsize)
-
-    def _setup_tds8(self):
-        """Wrap the TCP socket in TLS for TDS 8.0 strict encryption."""
-        LOG.debug("(TDS8) Setting up TDS 8.0 strict encryption")
-        context = ssl.SSLContext()
-        context.set_ciphers('ALL:@SECLEVEL=0')
-        context.minimum_version = ssl.TLSVersion.MINIMUM_SUPPORTED
-        context.check_hostname = False
-        context.verify_mode = ssl.CERT_NONE
-        context.set_alpn_protocols(["tds/8.0"])
-        self.socket = context.wrap_socket(self.socket, server_hostname=self.server)
-        self.tds8 = True
-        self.packetSize = 16 * 1024 - 1
-        LOG.info("(TDS8) TDS 8.0 TLS connection established")
 
     def initConnection(self):
-        #This is copied from tds.py
         self.connect()
 
         # Use a short timeout for the initial preLogin — if the server requires
@@ -113,57 +80,12 @@ class MYMSSQL(MSSQL):
             # With ENCRYPT_OFF, TLS is dropped after the first login packet (handled in sendNegotiate).
             # With ENCRYPT_REQ/ENCRYPT_ON, TLS stays active for the entire session.
             LOG.info("(MSSQL) Encryption required, switching to TLS")
-            # Creates a TLS context
-            context = ssl.SSLContext()
-            context.set_ciphers('ALL:@SECLEVEL=0')
-            context.minimum_version = ssl.TLSVersion.MINIMUM_SUPPORTED
-            context.verify_mode = ssl.CERT_NONE
-
-            # Here comes the important part, MSSQL server does not expect a raw TLS socket
-            # Instead it expects TDS packets to be sent in which TLS data is embedded
-            # Something like TDS_PACKET["Data"] = TLS_ENCRYPTED(data)
-            # To setup such a TLS tunnel inside another program, we need to use a STARTTLS like mechanism
-            # Which relies on MemoryBIO that are used to send data to the TLS context and receive data from it as well
-            # IN_BIO is where we send data to be encrypted and sent to the MSSQL server
-            in_bio = ssl.MemoryBIO()
-            # OUT_BIO is where we read data sent by the MSSQL server inside a TDS packet
-            out_bio = ssl.MemoryBIO()
-
-            # Now we can create the TLS object that will be used to manage handshake and data processing
-            tls = context.wrap_bio(in_bio, out_bio)
-
-            # So first let's handshake with the remote MSSQL server
-            while True:
-                try:
-                    tls.do_handshake()
-                except ssl.SSLWantReadError:
-                    # If we get a SSLWantReadError then it means the server received enough data and want to send some to us
-                    # So we read the data sent by the server and we send it back to it inside a TDS_PRE_LOGIN packet
-                    # That's the actual TLS server hello
-                    data = out_bio.read(4096)
-                    self.sendTDS(TDS_PRE_LOGIN, data, 0)
-
-                    # Now we read data one more time to extract the final TLS message
-                    tds_packet = self.recvTDS(4096)
-                    tls_data = tds_packet["Data"]
-
-                    # And we send that data to the in_bio object to complete the handshake
-                    in_bio.write(tls_data)
-                else:
-                    break
-
-            # At this point the TLS context is set up so we just store object inside the MSSQL class
-            # That will be used to encrypt/decrypt data and send them to the MSSQL server
-            self.packetSize = 16 * 1024 - 1
-            self.tlsSocket = tls
-            self.in_bio = in_bio
-            self.out_bio = out_bio
+            self.set_tls_context()
 
         self.resp = resp
         return True
 
     def sendNegotiate(self, negotiateMessage):
-        #Also partly copied from tds.py
         login = TDS_LOGIN()
 
         login['HostName'] = (''.join([random.choice(string.ascii_letters) for _ in range(8)])).encode('utf-16le')
@@ -206,46 +128,11 @@ class MYMSSQL(MSSQL):
         tds = self.recvTDS()
         self.replies = self.parseReply(tds['Data'])
         if TDS_LOGINACK_TOKEN in self.replies:
-            #Once we are here, there is a full connection and we can
-            #do whatever the current user has rights to do
             self.sessionData['AUTH_ANSWER'] = tds
             return None, STATUS_SUCCESS
         else:
             self.printReplies()
             return None, STATUS_ACCESS_DENIED
-
-    def sql_query(self, cmd, tuplemode=False, wait=True):
-        return self.batch(cmd, tuplemode, wait)
-
-    def batch(self, cmd, tuplemode=False, wait=True):
-        """Override batch to add ALL_HEADERS for TDS 8.0."""
-        if self.tds8:
-            from impacket.tds import TDS_SQL_BATCH
-
-            self.rows = []
-            self.colMeta = []
-            self.lastError = False
-
-            # ALL_HEADERS with transaction descriptor (required by TDS 8.0)
-            all_headers = struct.pack('<I', 22)   # Total length of ALL_HEADERS
-            all_headers += struct.pack('<I', 18)  # Length of this header
-            all_headers += struct.pack('<H', 2)   # Header type: Transaction Descriptor
-            all_headers += struct.pack('<Q', 0)   # Transaction descriptor (no transaction)
-            all_headers += struct.pack('<I', 1)   # Outstanding request count
-
-            sql_text = (cmd + "\r\n").encode("utf-16le")
-            packet_data = all_headers + sql_text
-
-            self.sendTDS(TDS_SQL_BATCH, packet_data)
-
-            if wait:
-                tds = self.recvTDS()
-                self.replies = self.parseReply(tds["Data"], tuplemode)
-                return self.rows
-            else:
-                return True
-        else:
-            return super().batch(cmd, tuplemode, wait)
 
     def close(self):
         return self.disconnect()
